test(liquidation): factor out driver for re-use
